Move parse url tests to test_registry
